refactor: separate test files
